+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # gradle-nunit-plugin changelog
 
 ## 1.6
+### Changed
+* Default working directory of NUnit v3 to build\nunit\
 
 ## 1.5
 ### Fixed