+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # gradle-nunit-plugin changelog
 
+## 1.7
+### Added
+* `resultFormat` is added to NUnit task. It allows to set the test report format (nunit2 or nunit3) for NUnit v3.
+
 ## 1.6
 ### Added
 * `getCommandArgs` is added to NUnit task